+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `wait_completion` param on `get_performances`, `list_task_output_assets` and `get_task_output_asset` to block execution until execution is over ([#368](https://github.com/Substra/substra/pull/368))
 - `list_task_output_assets` and `get_task_output_asset` wait that the compute task is over before getting assets ([#369](https://github.com/Substra/substra/pull/369))
 
+### Changed
+
+- change how API responses are parsed to match server changes ([#379](https://github.com/Substra/substra/pull/379))
+
 ### Fixed
 
 - typo in exception ([#377](https://github.com/Substra/substra/pull/377))